requirements.txt
click==8.1.7
python-dotenv==1.0.0
requests==2.31.0